+++ pkgs/jnigen/java/src/main/java/com/github/dart_lang/jnigen/apisummarizer/elements/KotlinConstructor.java
@@ -6,21 +6,19 @@ package com.github.dart_lang.jnigen.apisummarizer.elements;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kotlinx.metadata.KmConstructor;
-import kotlinx.metadata.jvm.JvmExtensionsKt;
+import kotlin.metadata.KmConstructor;
+import kotlin.metadata.jvm.JvmExtensionsKt;
 
 public class KotlinConstructor {
   public String name;
   public String descriptor;
   public List<KotlinValueParameter> valueParameters;
-  public int flags;
 
   public static KotlinConstructor fromKmConstructor(KmConstructor c) {
     var ctor = new KotlinConstructor();
-    ctor.flags = c.getFlags();
     var signature = JvmExtensionsKt.getSignature(c);
     ctor.name = signature == null ? null : signature.getName();
-    ctor.descriptor = signature == null ? null : signature.getDesc();
+    ctor.descriptor = signature == null ? null : signature.getDescriptor();
     ctor.valueParameters =
         c.getValueParameters().stream()
             .map(KotlinValueParameter::fromKmValueParameter)